+++ workers/unity/Packages/io.improbable.gdk.mobile/Editor/MobileLaunchUtils.cs
@@ -1,41 +1,24 @@
-using System.Text;
-using Improbable.Gdk.Core;
-using Improbable.Gdk.Tools;
-using UnityEngine;
+using System;
 
 namespace Improbable.Gdk.Mobile
 {
-    public class MobileLaunchUtils
+    public static class MobileLaunchUtils
     {
-        private const string argStructure = "+{0} {1} ";
-
-        public static string PrepareArguments(bool shouldConnectLocally, string runtimeIp)
+        public static string ToPrettyDeviceType(this DeviceType deviceType)
         {
-            var arguments = new StringBuilder();
-            if (shouldConnectLocally)
+            switch (deviceType)
             {
-                if (string.IsNullOrEmpty(runtimeIp))
-                {
-                    Debug.LogWarning("No local runtime IP was specified. Ensure you set one in SpatialOS > GDK tools configuration.");
-                }
-
-                arguments.AppendFormat(argStructure, RuntimeConfigNames.Environment, RuntimeConfigDefaults.LocalEnvironment);
-                arguments.AppendFormat(argStructure, RuntimeConfigNames.ReceptionistHost, runtimeIp);
+                case DeviceType.AndroidDevice:
+                    return "Android Device";
+                case DeviceType.AndroidEmulator:
+                    return "Android Emulator";
+                case DeviceType.iOSDevice:
+                    return "iOS Device";
+                case DeviceType.iOSSimulator:
+                    return "iOS Simulator";
+                default:
+                    throw new ArgumentException($"Unknown Device Type: {deviceType}");
             }
-            else
-            {
-                arguments.AppendFormat(argStructure, RuntimeConfigNames.Environment, RuntimeConfigDefaults.CloudEnvironment);
-
-                // Return error if no DevAuthToken is set AND fails to generate new DevAuthToken.
-                if (!PlayerPrefs.HasKey(RuntimeConfigNames.DevAuthTokenKey) && !DevAuthTokenUtils.TryGenerate())
-                {
-                    throw new PlayerPrefsException("Failed to retrieve a Dev Auth Token to launch a mobile client.");
-                }
-
-                arguments.AppendFormat(argStructure, RuntimeConfigNames.DevAuthTokenKey, DevAuthTokenUtils.DevAuthToken);
-            }
-
-            return arguments.ToString();
         }
     }
 }