+++ workers/unity/Packages/io.improbable.gdk.mobile/Editor/MobileLauncherWindow.cs
@@ -1,3 +1,5 @@
+using System;
+using System.Collections.Generic;
 using System.Linq;
 using Improbable.Gdk.Core.Editor;
 using UnityEditor;
@@ -17,13 +19,15 @@ namespace Improbable.Gdk.Mobile
         private static readonly Color DarkGrey = new Color(0.4f, 0.4f, 0.4f, 1);
         private static readonly Color LightGrey = new Color(0.7f, 0.7f, 0.7f, 1);
 
-        private MobileLaunchConfig launchConfig;
+        private MobileLaunchConfig mobileLaunchConfig;
 
-        private string[] simulatorNames;
-        private string[] deviceNames;
+        private (int currentIndex, List<DeviceLaunchConfig> devices) androidEmulators;
+        private (int currentIndex, List<DeviceLaunchConfig> devices) androidDevices;
 
-        private int simulatorNameIndex;
-        private int deviceNameIndex;
+        private (int currentIndex, List<DeviceLaunchConfig> devices) iOSSimulators;
+        private (int currentIndex, List<DeviceLaunchConfig> devices) iOSDevices;
+
+        private readonly string[] emptyDeviceNameList = { "No devices found" };
 
         [MenuItem("SpatialOS/Mobile Launcher", false, 52)]
         public static void ShowWindow()
@@ -37,9 +41,10 @@ namespace Improbable.Gdk.Mobile
 
         private void OnEnable()
         {
+            RefreshAndroidEmulatorsAndDevices();
+
 #if UNITY_EDITOR_OSX
-            deviceNames = iOSLaunchUtils.RetrieveAvailableiOSDevices().Keys.ToArray();
-            simulatorNames = iOSLaunchUtils.RetrieveAvailableiOSSimulators().Keys.ToArray();
+            RefreshiOSEmulatorsAndDevices();
 #endif
         }
 
@@ -48,112 +53,174 @@ namespace Improbable.Gdk.Mobile
             GUILayout.Label(MobileSectionLabel, EditorStyles.boldLabel);
             using (new EditorGUI.IndentLevelScope())
             {
-                launchConfig.RuntimeIp = EditorGUILayout.TextField(RuntimeIpLabel, launchConfig.RuntimeIp);
-                launchConfig.ShouldConnectLocally = EditorGUILayout.Toggle(ConnectLocallyLabel, launchConfig.ShouldConnectLocally);
+                mobileLaunchConfig.RuntimeIp = EditorGUILayout.TextField(RuntimeIpLabel, mobileLaunchConfig.RuntimeIp);
+                mobileLaunchConfig.ShouldConnectLocally = EditorGUILayout.Toggle(ConnectLocallyLabel, mobileLaunchConfig.ShouldConnectLocally);
             }
 
-            CommonUIElements.DrawHorizontalLine(10, LightGrey);
-            GUILayout.Label(AndroidSectionLabel, EditorStyles.boldLabel);
-
-            if (GUILayout.Button("Launch Android app"))
-            {
-                AndroidLaunchUtils.Launch(launchConfig.ShouldConnectLocally, launchConfig.RuntimeIp);
-            }
+            DisplayAndroidMenu();
 
 #if UNITY_EDITOR_OSX
+            DisplayiOSMenu();
+#endif
+        }
+
+        private void RefreshEmulatorsAndDevices(DeviceType deviceType)
+        {
+            switch (deviceType)
+            {
+                case DeviceType.AndroidDevice:
+                case DeviceType.AndroidEmulator:
+                    RefreshAndroidEmulatorsAndDevices();
+                    break;
+                case DeviceType.iOSDevice:
+                case DeviceType.iOSSimulator:
+                    RefreshiOSEmulatorsAndDevices();
+                    break;
+                default:
+                    Debug.LogError($"Unknown Device Type: {deviceType}");
+                    break;
+            }
+        }
+
+        private void RefreshAndroidEmulatorsAndDevices()
+        {
+            var (emulators, devices) = AndroidLaunchUtils.RetrieveAvailableEmulatorsAndDevices();
+
+            androidEmulators = (0, emulators);
+            androidDevices = (0, devices);
+        }
+
+        private void RefreshiOSEmulatorsAndDevices()
+        {
+            var (emulators, devices) = iOSLaunchUtils.RetrieveAvailableEmulatorsAndDevices();
+
+            iOSSimulators = (0, emulators);
+            iOSDevices = (0, devices);
+        }
+
+        private void DisplayAndroidMenu()
+        {
+            CommonUIElements.DrawHorizontalLine(10, LightGrey);
+
+            GUILayout.Label(AndroidSectionLabel, EditorStyles.boldLabel);
+
+            using (new EditorGUI.IndentLevelScope())
+            {
+                DisplayDevicesAndLaunchButton(DeviceType.AndroidEmulator);
+
+                CommonUIElements.DrawHorizontalLine(8, DarkGrey);
+
+                DisplayDevicesAndLaunchButton(DeviceType.AndroidDevice);
+            }
+        }
+
+#if UNITY_EDITOR_OSX
+        private void DisplayiOSMenu()
+        {
             CommonUIElements.DrawHorizontalLine(10, LightGrey);
 
             GUILayout.Label(iOSSectionLabel, EditorStyles.boldLabel);
 
             using (new EditorGUI.IndentLevelScope())
             {
-                launchConfig.DevelopmentTeamId = EditorGUILayout.TextField(DevelopmentTeamIdLabel, launchConfig.DevelopmentTeamId);
+                mobileLaunchConfig.DevelopmentTeamId = EditorGUILayout.TextField(DevelopmentTeamIdLabel, mobileLaunchConfig.DevelopmentTeamId);
 
                 if (GUILayout.Button("Build XCode project"))
                 {
-                    iOSLaunchUtils.Build(launchConfig.DevelopmentTeamId);
+                    iOSLaunchUtils.Build(mobileLaunchConfig.DevelopmentTeamId);
                 }
 
                 CommonUIElements.DrawHorizontalLine(8, DarkGrey);
 
-                using (new GUILayout.HorizontalScope())
-                {
-                    using (new EditorGUI.DisabledScope(simulatorNames.Length == 0))
-                    {
-                        simulatorNameIndex =
-                            EditorGUILayout.Popup("Simulator Model", simulatorNameIndex, simulatorNames);
-                    }
-
-                    var buttonIcon = new GUIContent(EditorGUIUtility.IconContent("Refresh"))
-                    {
-                        tooltip = "Refresh your simulator list."
-                    };
-
-                    if (GUILayout.Button(buttonIcon, EditorStyles.miniButton, GUILayout.ExpandWidth(false)))
-                    {
-                        simulatorNames = iOSLaunchUtils.RetrieveAvailableiOSSimulators().Keys.ToArray();
-                        simulatorNameIndex = 0;
-                    }
-                }
-
-                using (new EditorGUI.DisabledScope(simulatorNames.Length == 0))
-                {
-                    if (GUILayout.Button("Launch iOS app in Simulator"))
-                    {
-                        var availableSimulators = iOSLaunchUtils.RetrieveAvailableiOSSimulators();
-                        if (availableSimulators.TryGetValue(simulatorNames[simulatorNameIndex], out var simulatorUID))
-                        {
-                            iOSLaunchUtils.Launch(launchConfig.ShouldConnectLocally, simulatorUID, launchConfig.RuntimeIp, true);
-                        }
-                        else
-                        {
-                            simulatorNames = availableSimulators.Keys.ToArray();
-                            simulatorNameIndex = 0;
-                            Debug.LogError("Failed to launch app on selected simulator.");
-                        }
-                    }
-                }
+                DisplayDevicesAndLaunchButton(DeviceType.iOSSimulator);
 
                 CommonUIElements.DrawHorizontalLine(8, DarkGrey);
 
-                using (new GUILayout.HorizontalScope())
+                DisplayDevicesAndLaunchButton(DeviceType.iOSDevice);
+            }
+        }
+#endif
+
+        private void DisplayDevicesAndLaunchButton(DeviceType deviceType)
+        {
+            int index;
+            List<DeviceLaunchConfig> devices;
+
+            switch (deviceType)
+            {
+                case DeviceType.AndroidDevice:
+                    (index, devices) = androidDevices;
+                    break;
+                case DeviceType.AndroidEmulator:
+                    (index, devices) = androidEmulators;
+                    break;
+                case DeviceType.iOSDevice:
+                    (index, devices) = iOSDevices;
+                    break;
+                case DeviceType.iOSSimulator:
+                    (index, devices) = iOSSimulators;
+                    break;
+                default:
+                    throw new ArgumentException($"Unknown Device Type: {deviceType}");
+            }
+
+            var prettyDeviceType = deviceType.ToPrettyDeviceType();
+            var names = devices.Select(config => config.DeviceName).ToArray();
+            var noDevicesFound = devices.Count == 0;
+
+            // List of devices/emulators with refresh button linked to correct platform
+            using (new GUILayout.HorizontalScope())
+            {
+                using (new EditorGUI.DisabledScope(noDevicesFound))
                 {
-                    using (new EditorGUI.DisabledScope(deviceNames.Length == 0))
-                    {
-                        deviceNameIndex = EditorGUILayout.Popup("Device Model", deviceNameIndex, deviceNames);
-                    }
-
-                    var buttonIcon = new GUIContent(EditorGUIUtility.IconContent("Refresh"))
-                    {
-                        tooltip = "Refresh your device list."
-                    };
-
-                    if (GUILayout.Button(buttonIcon, EditorStyles.miniButton, GUILayout.ExpandWidth(false)))
-                    {
-                        deviceNames = iOSLaunchUtils.RetrieveAvailableiOSDevices().Keys.ToArray();
-                        deviceNameIndex = 0;
-                    }
+                    SetDeviceIndex(deviceType,
+                        noDevicesFound
+                            ? EditorGUILayout.Popup(prettyDeviceType, 0, emptyDeviceNameList)
+                            : EditorGUILayout.Popup(prettyDeviceType, index, names));
                 }
 
-                using (new EditorGUI.DisabledScope(deviceNames.Length == 0))
+                var buttonIcon = new GUIContent(EditorGUIUtility.IconContent("Refresh"))
                 {
-                    if (GUILayout.Button("Launch iOS app on device"))
-                    {
-                        var availableDevices = iOSLaunchUtils.RetrieveAvailableiOSDevices();
-                        if (availableDevices.TryGetValue(deviceNames[deviceNameIndex], out var deviceUID))
-                        {
-                            iOSLaunchUtils.Launch(launchConfig.ShouldConnectLocally, deviceUID, launchConfig.RuntimeIp, false);
-                        }
-                        else
-                        {
-                            deviceNames = availableDevices.Keys.ToArray();
-                            deviceNameIndex = 0;
-                            Debug.LogError("Failed to launch app on selected device. Is the device still connected?");
-                        }
-                    }
+                    tooltip = $"Refresh your {prettyDeviceType} list."
+                };
+
+                if (GUILayout.Button(buttonIcon, EditorStyles.miniButton, GUILayout.ExpandWidth(false)))
+                {
+                    RefreshEmulatorsAndDevices(deviceType);
                 }
             }
-#endif
+
+            // Draw a launch button that will launch the app on the correct platform and chosen device
+            using (new EditorGUI.DisabledScope(noDevicesFound))
+            {
+                if (!GUILayout.Button($"Launch app on {prettyDeviceType}"))
+                {
+                    return;
+                }
+
+                devices[index].Launch(mobileLaunchConfig);
+            }
+        }
+
+        private void SetDeviceIndex(DeviceType deviceType, int index)
+        {
+            switch (deviceType)
+            {
+                case DeviceType.AndroidDevice:
+                    androidDevices.currentIndex = index;
+                    break;
+                case DeviceType.AndroidEmulator:
+                    androidEmulators.currentIndex = index;
+                    break;
+                case DeviceType.iOSDevice:
+                    iOSDevices.currentIndex = index;
+                    break;
+                case DeviceType.iOSSimulator:
+                    iOSSimulators.currentIndex = index;
+                    break;
+                default:
+                    throw new ArgumentException($"Unknown Device Type: {deviceType}");
+            }
         }
     }
 }